complete series

+++ app/Http/Controllers/API/UserController.php
@@ -9,6 +9,11 @@ use App\User;
 
 class UserController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth:api');
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -16,7 +21,10 @@ class UserController extends Controller
      */
     public function index()
     {
-        return User::latest()->paginate(10);
+        if(\Gate::allows('isAdmin') || \Gate::allows('isAuthor')){
+
+            return User::latest()->paginate(2);
+        }
     }
 
     /**
@@ -40,6 +48,60 @@ class UserController extends Controller
             // 'photo' => $request['photo'],
             'password' => Hash::make($request['password']),
         ]);
+    }
+
+    /**
+     * Display the specified resource.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function updateProfile(Request $request)
+    {
+        $user = auth('api')->user();
+
+        $this->validate($request,[
+            'name' => 'required|string|max:191',
+            'email' => 'required|string|email|max:191|unique:users,email,'.$user->id,
+            'password' => 'sometimes|required|min:6'
+        ]);
+
+        $currentPhoto = $user->photo;
+
+        if($request->photo != $currentPhoto) {
+
+            $name = time().'.' . explode('/', explode(':', substr($request->photo, 0, strpos($request->photo, ';')))[1])[1];
+
+            \Image::make($request->photo)->save(public_path('img/profile/').$name);
+
+            $request->merge(['photo' => $name]);
+
+            $userPhoto = public_path('img/profile/').$currentPhoto;
+            
+            if(file_exists($userPhoto)){
+                @unlink($userPhoto);
+            }
+        }
+
+        if(!empty($request->password)){
+            $request->merge(['password' => Hash::make($request['password'])]);
+
+        }
+        
+        $user->update($request->all());
+
+        return ['message' => 'Success!'];
+    }
+    
+    /**
+     * Display the specified resource.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function profile()
+    {
+        return auth('api')->user();
     }
 
     /**
@@ -83,10 +145,34 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('isAdmin');
+        
         $user = User::findOrFail($id);
 
-        $user->delete();
+        if($user->id == auth()->user()->id){
+            return back()->with(['message' => 'User Deleted!']);
+        }else{
+            $user->delete();
+            return ['message' => 'User Deleted!'];
+        }
 
-        return ['message' => 'User Deleted!'];
     }
+
+    public function search(){
+
+        if ($search = \Request::get('q')) {
+            $users = User::where(function($query) use ($search){
+                $query->where('name','LIKE',"%$search%")
+                        ->orWhere('email','LIKE',"%$search%")
+                        ->orWhere('type','LIKE',"%$search%");
+            })->paginate(2);
+        }else{
+            $users = User::latest()->paginate(2);
+        }
+
+        return $users;
+
+    }
+
+
 }
